gWidgets-classes {gWidgets}	R Documentation

Classes for gWidgets instances

Description

Classes for gWidgets objects

Details

The gWidgets API is designed to have a double dispatch. The gWidgets package calling methods in the gWidgetsXXX packages. As such, the real class structure sits inside the toolkit packages. These classes are merely here for organization and some day documentation.

The "ANY" classes are a means to write compound widgets in gWidgets that work across the toolkits.

gAction gButton gCalendar gCheckbox gCheckboxGroup gCombobox gCommandline gDf gDfNotebook gEdit gExpandGroup gFilebrowse gFormLayout gFrame gGenericWidget gGraphics gGraphicsNotebook gGroup gHelp gHelpBrowser gHtml gImage gLabel gLayout gMenu gNotebook gPanedGroup gRadio gSeparator gSlider gSpinbutton gStatusbar gSvg gTable gText gToolbar gTree guiComponentRangeSelector guiComponentWithItems gVarBrowser gWindow

gCommandlineANY gGenericWidgetANY gFormLayoutANY gHelpbrowserANY gHelpANY

[Package gWidgets version 0.0-54.2 Index]