gcommandline {gWidgets}	R Documentation

A command line interface

Description

This constructs a simple command line interface for R

Usage

gcommandline(command = "", assignto = NULL, useGUI = TRUE, 
useConsole = FALSE, prompt = getOption("prompt"), width = 500, 
height = 0.6 * width, container = NULL, ...,toolkit =
                 guiToolkit())

Arguments

command
Initial command to evalues

assignto
Assigns output to this variable is non-NULL

useGUI
Is result also printed to GUI. Use FALSE to get text-only instance

useConsole
Is result also printed to console?

prompt
Prompt to use

width
Width of widget in pixels

height
Height of widget in pixels

container
Optional container to attach to

...
Ignored for now

toolkit
Which GUI toolkit to use

Details

Additional commands can be added programmatically with the svalue<- method. The The value assigned is a string containing the command. If it has a names attribute, this is taken as the variable name to assign the output to.

The svalue method returns the command history.

The "[" method can be used to retrieve the command history as well.

Examples

## Not run: 
  obj = gcommandline(container=TRUE)
  svalue(obj) <- "2+2"
  ## assign to x
  command = "rnorm(100)"; names(command) = "x"
  svalue(obj) <- command
  ## look at history
  obj[]

## End(Not run)

[Package gWidgets version 0.0-54.2 Index]