ghtml {gWidgets}	R Documentation

Constructors for widgets to handle text input

Description

The ghtml widget is intended to show HTML text either from a url or from a character string. Currently no toolkits support this widget, although it is in gWidgetsWWW.

Usage

ghtml(x, handler = NULL, action = NULL, container = NULL, ..., toolkit = guiToolkit())

Arguments

x
url or HTML-marked up character string to load into widget.

handler
(In theory, not implemented yet) Handles a click on a URL. The default is to open the clicked url in the widget. To override, the first argument, a list h, has component h\$url containing the url.

action
Passed along to the handler as h\$action

container
Optional container to attach widget to

...
Passed to add method of container

toolkit
Which GUI toolkit to use

Details

This widget loads the given url into a widget. Currently no toolkits support this.

The svalue method returns the current url or character string.

The svalue<- method loads a url or character string in the widget.

Examples

## Not run: 
  ghtml(system.file("html","gedit.html",package="gWidgets"),
    container = gwindow())

## End(Not run)

[Package gWidgets version 0.0-54.2 Index]