gpanedgroup {gWidgets}	R Documentation

A paned group holds two child components with a handle, or sash, between them to adjust the amount of space allocated to each

Description

A constructor for a paned group.

Usage

gpanedgroup(widget1=NULL, widget2=NULL, horizontal = TRUE, 
container = NULL, ..., toolkit = guiToolkit())

Arguments

widget1
Left (top) widget. Can be added at time of construction, or the add method can be used to add the child widgets one at a time.

widget2
Right (bottom) widget

horizontal
Left/right (TRUE) or top/bottom

container
Optional container to attach widget to

...
Passed to add method of container

toolkit
Which GUI toolkit to use

Details

The add method can be used to one child at a time.

The svalue method returns the sash position with a value between 0 and 1.

The svalue<- method can be used to specify the sash position with a value between 0 and 1.

Examples

## Not run: 
  w <- gwindow("gpanedgroup example")
  pg <- gpanedgroup(container=w)
  gvarbrowser(container = pg) ## first is left/top
  gtext(container = pg)
  svalue(pg) <- 0.25

## End(Not run)

[Package gWidgets version 0.0-54.2 Index]