gstatusbar {gWidgets}	R Documentation

Constructor of status bar widget

Description

A status bar widget is used to send message to the user. A familiar instance is the bottom area of a web browser.

Usage

gstatusbar(text = "", container = NULL, ..., toolkit = guiToolkit())

Arguments

text
Initial text of status bar

container
Optional container to attach widget to. Should be gwindow object

...
Passed to add method of container

toolkit
Which GUI toolkit to use

Details

The status bar simply shows a message in a label, typically at the bottom of a window. The svalue and svalue<- methods can be used to query or set the text. message onto the stack.

Statusbars should be added to the top-level gwindow instance.

Examples

## Not run: 
  w <- gwindow("status bar example")
  tbl <- list(quit=list(icon="quit",
    handler = function(...) dispose(w)))
  tb <- gtoolbar(tbl, container=w)
  sb <- gstatusbar("", container=w)
  txt <- gtext("type here", container=w)
  addHandlerChanged(txt, handler=function(h,...)
    svalue(sb) <- paste("You typed",svalue(txt),"in the  box",collapse=" "))

## End(Not run)

[Package gWidgets version 0.0-54.2 Index]