guiToolkit {gWidgets}	R Documentation

Function to select the GUI toolkit used by gWidgets

Description

A GUI toolkit is a separate package that implements the gWidgets API. This function allows one to select the toolkit to be used by default.

Usage

guiToolkit(name = NULL)

gtoolkit()

Arguments

name
The name matches the package name without the initial gWidgets. For instance, "RGtk2" refers to the implementation provided by gWidgetsRGtk2.

Details

Toolkits are named gWidgetsXXX. This function is used to get the toolkit class, or have the user select one. The gtoolkit function returns a string form of the toolkit. One way to set a toolkit is to use the guiToolkit option, as in options(guiToolkit="RGtk2").

Value

The guiToolkit function returns a subclass of guiWidgetsToolkit that is used for dispatching purposes by gWidgets. For example, the method svalue dispatches on its first argument and the value of the toolkit class stored in the toolkit slot of the object.

The gtoolkit function returns a character string for the toolkit.

Examples

	guiToolkit("RGtk2")

[Package gWidgets version 0.0-54.2 Index]