guiWidget-class {gWidgets}	R Documentation

Class "guiWidget" ~~~

Description

Base class for gWidget objects

Objects from the Class

Objects can be created by calls of the form new("guiWidget", ...).

These objects have two slots: a widget provided by a guiToolkit and a toolkit. Method dispatch is done on both values.

Slots

toolkit:

Object of class "guiWidgetsToolkit". A specification of which GUI toolkit the widget uses.

widget:

Object of class "ANY". A widget returned by the corresponding toolkit function.

Methods

.add

signature(obj = "guiWidget", toolkit = "guiWidgetsToolkitRGtk2", value = "ANY"): ...

.add

signature(obj = "guiWidget", toolkit = "guiWidgetsToolkitRGtk2", value = "guiWidgetORgWidgetRGtkORRGtkObject"): ...

.add

signature(obj = "gContainerRGtk", toolkit = "guiWidgetsToolkitRGtk2", value = "guiWidget"): ...

.add

signature(obj = "gMenuRGtk", toolkit = "guiWidgetsToolkitRGtk2", value = "guiWidget"): ...

.add

signature(obj = "gNotebookRGtk", toolkit = "guiWidgetsToolkitRGtk2", value = "guiWidget"): ...

.add

signature(obj = "gTextRGtk", toolkit = "guiWidgetsToolkitRGtk2", value = "guiWidget"): ...

.delete

signature(obj = "gContainerRGtk", toolkit = "guiWidgetsToolkitRGtk2", widget = "guiWidget"): ...

.delete

signature(obj = "RGtkObject", toolkit = "guiWidgetsToolkitRGtk2", widget = "guiWidget"): ...

.delete

signature(obj = "gMenuRGtk", toolkit = "guiWidgetsToolkitRGtk2", widget = "guiWidget"): ...

.svalue<-

signature(obj = "gMenuRGtk", toolkit = "guiWidgetsToolkitRGtk2", value = "guiWidget"): ...

.tag

signature(obj = "guiWidget", toolkit = "guiWidgetsToolkitRGtk2"): ...

.tag<-

signature(obj = "guiWidget", toolkit = "guiWidgetsToolkitRGtk2"): ...

[

signature(x = "guiWidget"): ...

[<-

signature(x = "guiWidget"): ...

add3rdmousepopupmenu

signature(obj = "guiWidget"): ...

add

signature(obj = "guiWidget"): ...

addSpace

signature(obj = "guiWidget"): ...

addSpring

signature(obj = "guiWidget"): ...

adddropmotion

signature(obj = "guiWidget"): ...

adddropsource

signature(obj = "guiWidget"): ...

adddroptarget

signature(obj = "guiWidget"): ...

addhandlerchanged

signature(obj = "guiWidget"): ...

addhandlerclicked

signature(obj = "guiWidget"): ...

addhandlerdestroy

signature(obj = "guiWidget"): ...

addhandlerdoubleclick

signature(obj = "guiWidget"): ...

addhandlerexpose

signature(obj = "guiWidget"): ...

addhandleridle

signature(obj = "guiWidget"): ...

addhandlerkeystroke

signature(obj = "guiWidget"): ...

addhandlerrightclick

signature(obj = "guiWidget"): ...

addhandlerunrealize

signature(obj = "guiWidget"): ...

addpopupmenu

signature(obj = "guiWidget"): ...

defaultWidget

signature(obj = "guiWidget"): ...

defaultWidget<-

signature(obj = "guiWidget"): ...

delete

signature(obj = "guiWidget"): ...

dim

signature(x = "guiWidget"): ...

dimnames

signature(x = "guiWidget"): ...

dimnames<-

signature(x = "guiWidget"): ...

dispose

signature(obj = "guiWidget"): ...

editable

signature(obj = "guiWidget"): ...

editable<-

signature(obj = "guiWidget"): ...

enabled

signature(obj = "guiWidget"): ...

enabled<-

signature(obj = "guiWidget"): ...

focus

signature(obj = "guiWidget"): ...

focus<-

signature(obj = "guiWidget"): ...

tooltip<-

signature(obj = "guiWidget"): ...

font

signature(obj = "guiWidget"): ...

font<-

signature(obj = "guiWidget"): ...

id

signature(obj = "guiWidget"): ...

id<-

signature(obj = "guiWidget"): ...

length

signature(x = "guiWidget"): ...

names

signature(x = "guiWidget"): ...

names<-

signature(x = "guiWidget"): ...

removehandler

signature(obj = "guiWidget"): ...

size

signature(obj = "guiWidget"): ...

size<-

signature(obj = "guiWidget"): ...

svalue

signature(obj = "guiWidget"): ...

svalue<-

signature(obj = "guiWidget"): ...

tag

signature(obj = "guiWidget"): ...

tag<-

signature(obj = "guiWidget"): ...

update

signature(object = "guiWidget"): ...

visible

signature(obj = "guiWidget"): ...

visible<-

signature(obj = "guiWidget"): ...

isExtant

signature(obj = "guiWidget"): ...

Note

Within gWidgets there are three main subclasses guiContainers, guiComponents and guiDialogs. The distinctions are more clear at the toolkit level.

Author(s)

John Verzani

[Package gWidgets version 0.0-54.2 Index]