gvarbrowser {gWidgets}	R Documentation

Widget for browsing environment

Description

A widget to browse the objects in the current global environment

Usage

gvarbrowser(handler = NULL,
                 action = "summary", container = NULL, ...,
                 toolkit = guiToolkit())

Arguments

handler
Handler for double click. Default is to call value of action on the object

action
Passed to handler.

container
Optional container to attach widget to

...
ignored

toolkit
Which GUI toolkit to use

Details

In gWidgetsRGtk2 there is an idle handler that updates the top-level components. However, changes are made below this, they will not be updated automatically. Rather, the user must close and expand that level.

The handler is called on the widget that provides the display, and not the gvarbrowser widget. If you want that in the handler, be sure to pass it in via the action argument.

svalue returns a character string containing the selected variable.

Examples

## Not run: 
gvarbrowser(container=TRUE)

## End(Not run)

[Package gWidgets version 0.0-54.2 Index]